node daemon can rus as windows service
