+++ src/kernel/node_daemon/Cargo.toml
@@ -24,6 +24,9 @@ num_cpus = "*"
 time = { version = "^0.3", features = ["formatting"] }
 sysinfo = "*"
 
+[target.'cfg(windows)'.dependencies]
+windows-service = "0.7"
+
 cyfs-warp = { path = "../../cyfs_gateway/cyfs-warp" }
 cyfs-gateway-lib = { path = "../../cyfs_gateway/cyfs-gateway-lib" }
 buckyos-kit = { path = "../../components/buckyos-kit" }
